Refactor package structure for poet extensions
