+++ components/generator/core/build.gradle.kts
@@ -9,6 +9,7 @@ kotlin {
         commonMain.dependencies {
             implementation(projects.sdk.ir.core)
             api(projects.sdk.core.tree)
+            implementation(projects.sdk.generator.kt.common.util)
         }
         commonTest.dependencies {
             implementation(libs.bundles.kmp.test)
