+++ sdk/generator/kt/common/poet-extensions/api/poet-extensions.api
@@ -1,4 +1,4 @@
-public final class io/github/composegears/valkyrie/generator/jvm/ext/CodeBlockKt {
+public final class io/github/composegears/valkyrie/sdk/generator/kt/poet/CodeBlockKt {
 	public static final fun argumentBlock (Lcom/squareup/kotlinpoet/CodeBlock$Builder;Ljava/lang/String;[Ljava/lang/Object;ZLkotlin/jvm/functions/Function1;)V
 	public static synthetic fun argumentBlock$default (Lcom/squareup/kotlinpoet/CodeBlock$Builder;Ljava/lang/String;[Ljava/lang/Object;ZLkotlin/jvm/functions/Function1;ILjava/lang/Object;)V
 	public static final fun builderBlock (Lcom/squareup/kotlinpoet/CodeBlock$Builder;Ljava/lang/String;[Ljava/lang/Object;Lkotlin/jvm/functions/Function1;)V
@@ -6,7 +6,7 @@ public final class io/github/composegears/valkyrie/generator/jvm/ext/CodeBlockKt
 	public static final fun trailingComma (Lcom/squareup/kotlinpoet/CodeBlock$Builder;)V
 }
 
-public final class io/github/composegears/valkyrie/generator/jvm/ext/SpecKt {
+public final class io/github/composegears/valkyrie/sdk/generator/kt/poet/SpecKt {
 	public static final fun fileSpecBuilder (Ljava/lang/String;Ljava/lang/String;Lkotlin/jvm/functions/Function1;)Lcom/squareup/kotlinpoet/FileSpec;
 	public static final fun funSpecBuilder (Ljava/lang/String;Lkotlin/jvm/functions/Function1;)Lcom/squareup/kotlinpoet/FunSpec;
 	public static final fun getterFunSpecBuilder (Lkotlin/jvm/functions/Function1;)Lcom/squareup/kotlinpoet/FunSpec;
@@ -16,7 +16,7 @@ public final class io/github/composegears/valkyrie/generator/jvm/ext/SpecKt {
 	public static final fun setIndent (Lcom/squareup/kotlinpoet/FileSpec$Builder;I)Lcom/squareup/kotlinpoet/FileSpec$Builder;
 }
 
-public final class io/github/composegears/valkyrie/generator/jvm/ext/TypeNameKt {
+public final class io/github/composegears/valkyrie/sdk/generator/kt/poet/TypeNameKt {
 	public static final fun nullable (Lcom/squareup/kotlinpoet/TypeName;)Lcom/squareup/kotlinpoet/TypeName;
 }
