+++ sdk/generator/kt/common/poet-extensions/src/main/kotlin/io/github/composegears/valkyrie/sdk/generator/kt/poet/CodeBlock.kt
@@ -1,4 +1,4 @@
-package io.github.composegears.valkyrie.generator.jvm.ext
+package io.github.composegears.valkyrie.sdk.generator.kt.poet
 
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.withIndent
