+++ settings.gradle.kts
@@ -73,7 +73,6 @@ include("tools:idea-plugin")
 include("components:generator:core")
 include("components:generator:kmp:imagevector")
 include("components:generator:iconpack")
-include("components:generator:jvm:poet-extensions")
 include("components:generator:jvm:imagevector")
 include("components:parser:jvm:svg")
 include("components:parser:jvm:xml")
@@ -89,6 +88,8 @@ include("sdk:figma:converter")
 include("sdk:core:extensions")
 include("sdk:core:tree")
 include("sdk:core:xml")
+include("sdk:generator:kt:common:util")
+include("sdk:generator:kt:common:poet-extensions")
 include("sdk:generator:xml")
 include("sdk:intellij:psi:iconpack")
 include("sdk:intellij:psi:imagevector")
